==> 99-01-Collect-PbaLinks.ipynb <==
# Collect Pink Bike Links

This notebook extracts article links from the Pink Bike RSS feed, Trailforks Website, and a 2025 prediction article file. The ultimate goal is to match articles to specific bike products.

import importlib
import Utils
importlib.reload(Utils)  # This will print "Utils.py is loaded" again
import json

#pip install -r requirements.txt

url_pbrss = 'https://www.pinkbike.com/pinkbike_xml_feed.php'
pbrss_links = Utils.get_pinkbike_article_links_list_from_rss(url_pbrss)

url_tf = 'https://www.trailforks.com/articles/'
pbtf_links = Utils.get_pinkbike_article_links_list_from_trailforks(url_tf)

pb_pred2025 = "https://www.pinkbike.com/news/dario-digiulios-2025-predictions.html"
pbtf_links.append(pb_pred2025)

real_links = list(set(pbrss_links + pbtf_links))

links = ['https://www.pinkbike.com/news/review-industry-nine-hydra-2-870-points-of-engagement-and-increased-durability.html?trk=rss',
'https://www.pinkbike.com/news/first-ride-rovals-control-world-cup-wheelset-uses-carbon-spokes-and-weighs-only-991-grams.html?trk=rss',
'https://www.pinkbike.com/news/review-vorsprung-telum-coil-shock.html?trk=rss', 'https://www.pinkbike.com/news/first-ride-race-face-turbine-flat-pedals.html?trk=rss']

len(real_links)

# save list to json on disk
with open('data/01-real-links.json', 'w') as f:
    json.dump(real_links, f)
